requirements.txt
alembic==1.13.2
Authlib==1.3.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
flask-socketio==5.3.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pycparser==2.22
requests==2.32.3
SQLAlchemy==2.0.31
python-barcode==0.14.0
typing_extensions==4.12.2
urllib3==2.2.2
fpdf==1.7.2
pillow==11.0.0
Werkzeug==3.0.3
WTForms==3.1.2